requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.1.0
asgiref==3.7.2
async-generator==1.10
async-timeout==4.0.3
attrs==22.1.0
Automat==22.10.0
beautifulsoup4==4.11.1
bio==1.5.9
biopython==1.81
biothings-client==0.3.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
chromedriver-autoinstaller==0.6.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==6.0.0
constantly==23.10.4
crochet==2.1.1
cryptography==37.0.4
cssselect==1.2.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.6
django-import-export==3.3.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
etelemetry==0.3.0
exceptiongroup==1.2.0
filelock==3.12.4
fpdf==1.7.2
frozenlist==1.4.1
future==0.18.3
googlesearch-python==1.2.3
gprofiler-official==1.0.0
h11==0.13.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.2
hyperlink==21.0.0
idna==3.3
incremental==22.10.0
isodate==0.6.1
itemadapter==0.8.0
itemloaders==1.1.0
Jinja2==3.1.2
jmespath==1.0.1
looseversion==1.3.0
lxml==4.9.3
MarkupPy==1.14
MarkupSafe==2.1.3
maxminddb==2.4.0
maxminddb-geolite2==2018.703
multidict==6.0.4
mygene==3.2.2
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
numpy==1.23.1
odfpy==1.4.1
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==1.4.3
parsel==1.8.1
pathlib==1.0.1
pdfkit==1.0.0
Pillow==10.0.1
platformdirs==3.10.0
pooch==1.7.0
Protego==0.3.0
prov==2.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pydot==1.4.2
Pygments==2.17.2
pymed==0.8.9
PyMuPDF==1.23.3
PyMuPDFb==1.23.3
pyOpenSSL==22.0.0
pyparsing==3.1.1
pypdf==3.16.1
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.2
pyxnat==1.6
PyYAML==6.0.1
queuelib==1.6.2
rdflib==7.0.0
reportlab==4.0.4
requests==2.28.1
requests-file==1.5.1
scipy==1.11.2
Scrapy==2.11.0
scrapydo==0.2.2
selenium==3.141.0
service-identity==23.1.0
simplejson==3.19.1
six==1.16.0
snakeviz==2.2.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.4
tablib==3.5.0
tldextract==5.1.1
tornado==6.4
tqdm==4.66.1
traits==6.3.2
trio==0.21.0
trio-websocket==0.9.2
Twisted==22.10.0
typing_extensions==4.8.0
urllib3==1.26.10
w3lib==2.1.2
webdriver-auto-update==1.0.0
wget==3.2
wrapt==1.16.0
wsproto==1.1.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.4
zope.interface==6.1
